chore: AvatarHUD as legacy
