--- unity-client/Assets/Scripts/MainScripts/DCL/Controllers/HUD/AvatarHUD/AvatarHUDModel.cs
@@ -1,9 +1,0 @@
-using UnityEngine;
-
-public class AvatarHUDModel
-{
-    public string name { get; set; } = "";
-    public string mail { get; set; } = "";
-    public Sprite avatarPic { get; set; } = null;
-    public int newWearables { get; set; } = 0;
-}
