+++ unity-client/Assets/Scripts/MainScripts/DCL/Controllers/HUD/AvatarHUD/Legacy.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c0afaa64f9a0a4ec0b459ad3556df388
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
