+++ unity-client/Assets/Scripts/MainScripts/DCL/Controllers/HUD/AvatarHUD/Legacy/AvatarHUDModel.cs
@@ -1,0 +1,12 @@
+using UnityEngine;
+
+namespace Legacy
+{
+    public class AvatarHUDModel
+    {
+        public string name { get; set; } = "";
+        public string mail { get; set; } = "";
+        public Sprite avatarPic { get; set; } = null;
+        public int newWearables { get; set; } = 0;
+    }
+}
