+++ unity-client/Assets/Scripts/MainScripts/DCL/Controllers/HUD/AvatarHUD/Tests/Legacy.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f0b0f55fc5ea340239355f29e63dda52
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
