+++ unity-client/Assets/Scripts/MainScripts/DCL/Controllers/HUD/AvatarHUD/Tests/Legacy/AvatarHUDTests.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 using UnityEngine.TestTools;
 using UnityEngine.UI;
 
-namespace Tests
+namespace Legacy.Tests
 {
     public class AvatarHUDTests : TestsBase
     {
